Move remaining fragments classes into android package. (#205)
