+++ CHANGELOG.md
@@ -30,6 +30,11 @@
 - Move `BackCallback` into `android` package.
 - Move `FragmentLifecycleEvent` into `android.events` package.
 - Move `Integration` into `android` package.
+- Move `FragmentLifecycleCallback` into `android` package.
+- Move `FragmentComponent` into `android` package.
+- Move `FragmentContract` into `android` package.
+- Move `FragmentFlowState` into `android` package.
+- Move `FragmentFlowStore` into `android` package.
 
 ## [0.6.1] - November 18, 2020
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.
