+++ formula-android/src/main/java/com/instacart/formula/android/ActivityStore.kt
@@ -1,8 +1,6 @@
 package com.instacart.formula.android
 
 import androidx.fragment.app.FragmentActivity
-import com.instacart.formula.fragment.FragmentFlowState
-import com.instacart.formula.fragment.FragmentFlowStore
 import com.instacart.formula.android.events.FragmentLifecycleEvent
 import io.reactivex.rxjava3.disposables.Disposable
