+++ formula-android/src/main/java/com/instacart/formula/android/ActivityStoreContext.kt
@@ -3,8 +3,6 @@ package com.instacart.formula.android
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.Lifecycle
 import com.instacart.formula.android.events.ActivityResult
-import com.instacart.formula.fragment.FragmentFlowState
-import com.instacart.formula.fragment.FragmentFlowStore
 import com.instacart.formula.android.events.FragmentLifecycleEvent
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.disposables.Disposable
