+++ formula-android/src/main/java/com/instacart/formula/android/Feature.kt
@@ -1,7 +1,6 @@
 package com.instacart.formula.android
 
 import io.reactivex.rxjava3.core.Observable
-import com.instacart.formula.fragment.FragmentFlowStore
 
 /**
  * Feature is based on uni-directional state management where a single state model drives
