+++ formula-android/src/main/java/com/instacart/formula/android/FeatureFactory.kt
@@ -1,7 +1,5 @@
 package com.instacart.formula.android
 
-import com.instacart.formula.fragment.FragmentFlowStore
-
 /**
  * Feature factory is responsible for creating a [Feature] for a specific [fragment key][Key].
  *
