+++ formula-android/src/test/java/com/instacart/formula/android/ActivityStoreContextTest.kt
@@ -3,7 +3,6 @@ package com.instacart.formula.android
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.Lifecycle
 import com.instacart.formula.android.internal.ActivityStoreContextImpl
-import com.instacart.formula.fragment.FragmentContract
 import com.jakewharton.rxrelay3.PublishRelay
 import com.nhaarman.mockito_kotlin.mock
 import com.nhaarman.mockito_kotlin.verify
